Make hawtio-dev-mode work regardless of location of hawtio clone
Change how the location of hawtio is recorded in the hawtio-dev-mode feature. Previously the clone was required to be called 'hawtio' and share a common parent with where fuse fabric was running. This is now changed to be the absolute path of where hawtio is built.

Also fixes a typo for Bundles display label: SymbolicName -> Symbolic Name