Remove the need for hawtio-osgi
hawtio-osgi used to add the OSGi manifest (created in hawtio-web) to the WAR file. However, an OSGi manifest can also be added to a plain WAR file making it both a WAR and an OSGi WAB at the same time.
Another issue with the previous setup was that because the OSGi manifest was generated in hawtio-web but applied in hawtio-osgi the identity (Bundle-SymbolicName) of the hawtio-osgi WAB was set to hawtio-web which is incorrect. This commit resolves that as well.