The beginnings of a OSGi Classloading Debugging MXBean
The MXBean can tell the client what bundle is providing the requested class given a classload operation in another bundle. This aids in debugging classloading problem with OSGi in that it can tell you why a ClassCastException occurs, or a ClassNotFoundException/NoClassDefFoundError...

Includes unit tests.
Note that the Bar.class is a class file that is deliberately checked in for testing purposes. It is used by the OSGiToolsTest class.
This commit helps implement #169.