Add hawtio-osgi-jmx to hawtio-web
This module provides a JMX bean to aid OSGi classloading debugging.
I guess it could be separated into a separate module but it's only 3kb so I'm not sure it's really necessary to isolate. The module can operate in both OSGi and non-OSGi contexts. It looks at whether OSGi is available. If it is, it registers the io.hawt.osgi.OSGiTools JMX Bean. If OSGi is not available the module does nothing.