Display Exported Package attributes and directived as popover, for #196
The exported packages in the single bundle view now have a popover that combines information from the manifest header with the information obtained from Package Admin.
The 'Version' is the actual version, as reported by Package Admin. The other directives and attributes are obtained from the Bundle Manifest are merged with this info, and nicely formatted.