Decorate imported packages with metadata from manifest headers. Fixes #196
This applies to the packages in the single bundle view.
The actual imported version is displayed along with additional metadata from the manifest.
Currently I'm using badges with the named-color scheme that goes with that. That can be changed though. Colours are used to differentiate between optional and required imports.
A required import shows as blue.
An optional import that is satisfied shows as grey.
An optional import that is not satisfied shows as orange.