add support for Bearer Token based Authentication
I tried to secure hawt.io console with keycloak based on hawt.io documentation but also based on the example of activemq artemis. Unfortunately i was not able to make it working and therefore started debugging. I found the problem in class io.hawt.system.Authenticator which does only support Basic Http Header Authorization but not the Bearer token. I have extended the class to make this working.
It seems that there is also a corresponding problem in the javascript code because even if i am redirected to keycloak authentication page the token is never passed with the request. FOr me so far this is no problem because i do authentication in a proxy service which will then pass the token properly the hawtio server.