fix(quarkus): HAWNG-429 fix loadFromHawtioStatic sanitisation
The previous fix tried to sanitise the path for /hawtconfig.json but the
path passed to loadFromHawtioStatic() is always constant thus forcing
sanitising it is redundant without no real gain. We should instead
clarify in the code that the path is always '/hawtconfig.json' to make
the analyzer tool satisfy.

The previous fix also caused the following warning in the log:
"Couldn't find file on location: file:///hawtconfig.json"